requirements.txt
# Core dependencies
openai>=1.12.0
pinecone-client>=3.0.0
numpy>=1.21.0
pandas>=1.3.0
python-dotenv>=0.19.0

# ML & NLP
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.0.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
urllib3>=2.0.0

# Development & testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=6.0.0
streamlit